+++ backend/app/utils.py
@@ -162,23 +162,3 @@ def process_query_result(response_data: List[Dict]) -> List[Dict[str, Any]]:
 
     # Convert defaultdict to list
     return list(users.values())
-
-  
-async def send_daily_email()->Any:
-    """
-    Email to inform user about unfinished todo_item
-    """
-    query_data = await get_user_with_todo_item_detail()
-    formatted_data = process_query_result(query_data)
-    
-    for user_data in formatted_data:
-        email_to = user_data["email"]
-        username = user_data["username"]
-        task_data = user_data["task_data"]
-    
-        email_data = generate_daily_status_mail(email_to=email_to,
-                                            username=username,
-                                            task_data=task_data,)
-        send_mail(email_to = email_to,
-                    subject=email_data.subject,
-                    html_content=email_data.html_content)
